+++ club_name_mapping.json
@@ -1,0 +1,73 @@
+{
+  "italia": [
+    {
+      "AC Milan": "AC Milan"
+    },
+    {
+      "Atalanta BC": "Atalanta"
+    },
+    {
+      "Benevento": "Benevento"
+    },
+    {
+      "Bologna FC 1909": "Bologna FC"
+    },
+    {
+      "Cagliari Calcio": "Cagliari"
+    },
+    {
+      "AC Chievo Verona": "Chievo"
+    },
+    {
+      "Crotone": "Crotone"
+    },
+    {
+      "ACF Fiorentina": "Fiorentina"
+    },
+    {
+      "Frosinone Calcio": "Frosinone"
+    },
+    {
+      "Genoa CFC": "Genoa"
+    },
+    {
+      "Hellas Verona": "Hellas Verona"
+    },
+    {
+      "FC Internazionale Milano": "Inter Milan"
+    },
+    {
+      "Juventus FC": "Juventus"
+    },
+    {
+      "SS Lazio": "Lazio"
+    },
+    {
+      "SSC Napoli": "Napoli"
+    },
+    {
+      "AS Roma": "AS Roma"
+    },
+    {
+      "Parma Calcio 1913": "Parma"
+    },
+    {
+      "SPAL 2013": "SPAL"
+    },
+    {
+      "UC Sampdoria": "Sampdoria"
+    },
+    {
+      "US Sassuolo Calcio": "Sassuolo"
+    },
+    {
+      "Torino FC": "Torino"
+    },
+    {
+      "Udinese Calcio": "Udinese"
+    },
+    {
+      "Empoli FC": "Empoli"
+    }
+  ]
+}
